server/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
Django==5.1.6
django-cors-headers==4.7.0
django-mssql-backend==2.8.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
idna==3.10
inflection==0.5.1
packaging==24.2
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
pyodbc==5.2.0
pyspnego==0.11.2
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
requests_ntlm==1.3.0
sqlparse==0.5.3
sspilib==0.2.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
